requirements.txt
slack-bolt==1.18.0
slack-sdk==3.21.3
SQLAlchemy==1.4.41
python-dotenv==1.0.0
psycopg2-binary==2.9.6
reportlab==4.0.4
requests==2.31.0
fpdf
openpyxl>=3.1.2
